requirements.txt
alembic==1.3.2
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.10.3
lxml==4.4.2
Mako==1.1.0
MarkupSafe==1.1.1
newsapi-python==0.2.6
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.18.0
pandas==0.25.3
Pillow==6.2.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-file==1.4.3
singledispatch==3.4.0.3
six==1.13.0
soupsieve==1.9.5
SQLAlchemy==1.3.12
tinysegmenter==0.3
tldextract==2.2.2
urllib3==1.25.7
Werkzeug==0.16.0
